backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
bcrypt==4.1.2
httpx==0.25.2
openai==1.6.1
emergent-integration==1.0.0
twilio==8.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
stripe==7.8.0
requests==2.31.0
email-validator==2.1.0
jinja2==3.1.2